requirements.txt
numpy~=1.19.5
pillow~=8.2.0
matplotlib~=3.3.4
scipy~=1.6.2
torchvision~=0.9.1
setuptools~=57.4.0
cython~=0.29.24
pysparkling~=0.6.1